refactor using a helper to ease testing
