+++ core/src/main/java/com/telefonica/euro_iaas/sdc/util/ConnectionSetup.java
@@ -1,0 +1,48 @@
+package com.telefonica.euro_iaas.sdc.util;
+
+import java.io.IOException;
+import java.net.URL;
+import java.security.KeyManagementException;
+import java.security.NoSuchAlgorithmException;
+import java.security.cert.X509Certificate;
+
+import javax.net.ssl.HostnameVerifier;
+import javax.net.ssl.HttpsURLConnection;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLSession;
+import javax.net.ssl.TrustManager;
+import javax.net.ssl.X509TrustManager;
+
+public class ConnectionSetup {
+
+    public HttpsURLConnection createConnection(URL url) throws NoSuchAlgorithmException, KeyManagementException,
+            IOException {
+        // Install the all-trusting trust manager
+        SSLContext sslctx = SSLContext.getInstance("SSL");
+        javax.net.ssl.TrustManager[] trustAllCerts = new TrustManager[] { new X509TrustManager() {
+            public X509Certificate[] getAcceptedIssuers() {
+                return null;
+            }
+
+            public void checkClientTrusted(X509Certificate[] certs, String authType) {
+            }
+
+            public void checkServerTrusted(X509Certificate[] certs, String authType) {
+            }
+        } };
+        sslctx.init(null, trustAllCerts, null);
+
+        HttpsURLConnection.setDefaultSSLSocketFactory(sslctx.getSocketFactory());
+        HostnameVerifier allHostsValid = new HostnameVerifier() {
+            public boolean verify(String hostname, SSLSession session) {
+                return true;
+            }
+        };
+        // Install the all-trusting host verifier
+        HttpsURLConnection.setDefaultHostnameVerifier(allHostsValid);
+        HttpsURLConnection con = (HttpsURLConnection) url.openConnection();
+
+        return con;
+
+    }
+}
